pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "academy-tutorial"
version = "0.1.0"
authors = [
    {name = "Alok Kamatar", email = "alokvk2@uchicago.edu"},
    {name = "Greg Pauloski"},
    {name = "Globus Labs"},
]
maintainers = [
    {name = "Alok Kamatar", email = "alokvk2@uchicago.edu"},
]
description = "Materials for academy tutorial"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
    "academy-py>=0.3.0",
    "emoji",
    "globus-compute-sdk"
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pre-commit",
    "ruff>=0.2.0",
]

[tool.codespell]
skip = """
.git,
.github,
__pycache__,
build,
dist,
.*egg-info
"""

[tool.mypy]
python_version = "3.12"
plugins = [
    "mypy.plugins.proper_plugin",
    "proxystore.mypy_plugin",
    "pydantic.mypy",
]
exclude = [
    'starter_code',
]
check_untyped_defs = true
disable_error_code = "import-untyped"
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "testing.*"
allow_incomplete_defs = true
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
allow_incomplete_defs = true
allow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"

[tool.ruff]
line-length = 79
target-version = "py39"

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"
quote-style = "single"

[tool.ruff.lint]
# See all rules here: https://beta.ruff.rs/docs/rules
select = [
    # pyflakes
    "F",
    # pycodestyle
    "E",
    # mccabe
    "C90",
    # isort
    "I",
    # pep8-naming
    "N",
    # pydocstyle
    "D",
    # pyupgrade
    "UP",
    # flake8-2020
    "YTT",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-commas
    "COM",
    # flake8-comprehensions
    "C4",
    # flake8-implicit-str-concat
    "ISC",
    # flake8-pytest-style
    "PT",
    # flake8-quotes
    "Q",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # PyLint
    "PL",
    # ruff-specific
    "RUF",
]
extend-ignore = [
    # Allow builtin module shadowing.
    "A005",
    # Do not require: module (D100), package (D104), magic method (D105),
    # or __init__ (D107) docstrings.
    "D100", "D104", "D105", "D107",
    # Allow nested with statements.
    "SIM117",
]

[tool.ruff.lint.flake8-pytest-style]
parametrize-values-type = "tuple"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "double"

[tool.ruff.lint.isort]
force-single-line = true
known-first-party = ["academy_tutorial"]
order-by-type = false
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"*/__init__.py" = ["F401"]
"tests/conftest.py" = ["F401"]
"*/*_test.py" = ["D10"]
"solutions/*" = ["D"]
"starter_code/*" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.setuptools.packages.find]
include = ["academy_tutorial"]
namespaces = false
